requirements.txt
openai==1.1.2
python-dotenv
flask
python-docx
pandas
streamlit
openpyxl
retrying